:recycle: refactor: refactor to client mode
