+++ src/client/index.ts
@@ -1,2 +1,4 @@
 export * from './const';
 export { fetchPluginMessage } from './message';
+export * from './type';
+export * from './useWatchPluginMessage';
