+++ src/client/utils.ts
@@ -1,6 +1,5 @@
-import { PluginRenderProps } from '@/types';
-
 import { PluginChannel } from './const';
+import { PluginRenderProps } from './type';
 
 export const onReceiveData = <T>(e: MessageEvent, onData: (data: PluginRenderProps<T>) => void) => {
   if (e.data.type === PluginChannel.renderPlugin) {
