--- src/react/index.ts
@@ -1,2 +1,0 @@
-export * from '../browser';
-export * from './useWatchPluginMessage';
