+++ src/types/index.ts
@@ -1,4 +1,3 @@
 export * from './chatGPT';
-export * from './client';
 export * from './manifest';
 export * from './market';
